DateTimePicker - allow migration to new the Dialog via migrateDialog
