+++ src/incubator/Dialog/types.ts
@@ -1,6 +1,7 @@
 import {PropsWithChildren, ReactElement} from 'react';
 import {StyleProp, TextStyle, ViewStyle} from 'react-native';
 import {AlignmentModifiers} from '../../commons/modifiers';
+import {DialogProps as DialogPropsOld} from '../../components/dialog';
 import {ModalProps} from '../../components/modal';
 import {ViewProps} from '../../components/view';
 import {TextProps} from '../../components/text';
@@ -122,3 +123,25 @@ export interface _DialogProps extends AlignmentModifiers, Pick<ViewProps, 'useSa
 }
 
 export type DialogProps = PropsWithChildren<_DialogProps>;
+
+// For migration purposes
+export interface _DialogPropsOld {
+  /**
+   * The props to pass to the dialog expandable container
+   */
+  dialogProps?: DialogPropsOld;
+  migrateDialog?: false;
+}
+
+export interface _DialogPropsNew {
+  /**
+   * The props to pass to the dialog expandable container
+   */
+  dialogProps?: DialogProps;
+  /**
+   * Migrate the Dialog to DialogNew (make sure you use only new props in dialogProps)
+   */
+  migrateDialog: true;
+}
+
+export type DialogMigrationProps = _DialogPropsOld | _DialogPropsNew;
